@@ -1,16 +1,114 @@
 import { NextResponse } from "next/server";
 import prisma from "@/lib/prisma";
-export  async function GET(
+import { createSupabaseServerClient } from "@/lib/supabase/server";
+
+export async function GET(
   request: Request,
   { params }: { params: { model: string } }
 ) {
   const model = params.model;
+  const supabase = createSupabaseServerClient();
+
+  // Attempt to get user and authError, but DO NOT immediately return 401.
+  // We will check for user/auth only when specific models require it.
+  const {
+    data: { user },
+    error: authError,
+  } = await supabase.auth.getUser();
+
+  let currentUserRole;
+  // Only attempt to fetch the user's role from DB if a user exists
+  if (user) {
+    try {
+      const dbuser = await prisma.user.findUnique({
+        where: {
+          id: user.id,
+        },
+        select: { role: true },
+      });
+      currentUserRole = dbuser?.role;
+    } catch (error) {
+      console.error("Error fetching user role from DB:", error);
+      // If there's a DB error fetching role, we can treat it as unauthenticated
+      // for public routes, but for restricted routes, it should still be an error.
+      // For now, let's re-throw or handle as a 500 for consistency in error handling.
+      return NextResponse.json(
+        { message: "Internal Server Error during role check" },
+        { status: 500 }
+      );
+    }
+  }
 
   try {
     let data;
 
     switch (model) {
-      case "users":
+      case "projects":
+        data = await prisma.project.findMany({
+          where: { is_active: true },
+          orderBy: { sort_order: "asc" },
+          include: {
+            creator: {
+              select: { name: true, email: true },
+            },
+          },
+        });
+               break;
+
+      case "services": // Existing: Open to all users
+        data = await prisma.service.findMany({
+          orderBy: { sort_order: "asc" },
+          where: { is_active: true }, // Assuming services are also public and active
+          include: {
+            creator: {
+              select: { name: true, email: true },
+            },
+          },
+        });
+        break;
+
+      case "testimonials": // Existing: Open to all users
+        data = await prisma.testimonial.findMany({
+          orderBy: { sort_order: "asc" },
+          where: { is_active: true }, // Assuming testimonials are also public and active
+          include: {
+            // creator: { select: { name: true, email: true } }, // Uncomment if Testimonial includes creator
+          },
+        });
+        break;
+
+      case "blog-posts": // Existing: Open to all users
+        data = await prisma.blogPost.findMany({
+          where: { is_published: true },
+          orderBy: { published_at: "desc" },
+          include: {
+            author: {
+              select: { name: true, email: true },
+            },
+          },
+        });
+        break;
+
+      case "team-members": // Existing: Open to all users
+        data = await prisma.teamMember.findMany({
+          orderBy: { sort_order: "asc" },
+          where: { is_active: true }, // Assuming team members are also public and active
+        });
+        break;
+
+      case "users": // Requires authentication and ADMIN role
+        if (authError || !user) {
+          return NextResponse.json(
+            { message: "Unauthorized" },
+            { status: 401 }
+          );
+        }
+        if (currentUserRole !== "ADMIN") {
+          return NextResponse.json(
+            { message: "Forbidden: Only Admins can view all users" },
+            { status: 403 }
+          );
+        }
         data = await prisma.user.findMany({
           select: {
             id: true,
@@ -24,57 +122,20 @@ export  async function GET(
           },
         });
         break;
-      case "services":
-        data = await prisma.service.findMany({
-          orderBy: { sort_order: "asc" },
-          include: {
-            creator: {
-              select: { name: true, email: true },
-            },
-          },
-        });
-        break;
-      case "portfolio-items":
-        data = await prisma.portfolioItem.findMany({
-          orderBy: { sort_order: "asc" },
-          include: {
-            service: {
-              select: { name: true, slug: true },
-            },
-            creator: {
-              select: { name: true, email: true },
-            },
-          },
-        });
-        break;
-      case "testimonials":
-        data = await prisma.testimonial.findMany({
-          orderBy: { sort_order: "asc" },
-          include: {
-            creator: {
-              select: { name: true, email: true },
-            },
-          },
-        });
-        break;
-      case "videos":
-        data = await prisma.video.findMany({
-          orderBy: { sort_order: "asc" },
-        });
-        break;
-      case "blog-posts":
-        data = await prisma.blogPost.findMany({
-          orderBy: { published_at: "desc" },
-          include: {
-            author: {
-              select: { name: true, email: true },
-            },
-          },
-        });
-        break;
-      case "inquiries":
-        // In a real application, you'd want to restrict access to inquiries
-        // to authenticated admin users. For now, we fetch all.
+
+      case "inquiries": // Requires authentication and ADMIN role
+        if (authError || !user) {
+          return NextResponse.json(
+            { message: "Unauthorized" },
+            { status: 401 }
+          );
+        }
+        if (currentUserRole !== "ADMIN") {
+          return NextResponse.json(
+            { message: "Forbidden: Access to inquiries restricted" },
+            { status: 403 }
+          );
+        }
         data = await prisma.inquiry.findMany({
           orderBy: { created_at: "desc" },
           include: {
@@ -82,33 +143,6 @@ export  async function GET(
               select: { name: true, slug: true },
             },
           },
-        });
-        break;
-      case "company-info":
-        data = await prisma.companyInfo.findFirst();
-        break;
-      case "team-members":
-        data = await prisma.teamMember.findMany({
-          orderBy: { sort_order: "asc" },
-        });
-        break;
-      case "faqs":
-        data = await prisma.fAQ.findMany({
-          orderBy: { sort_order: "asc" },
-        });
-        break;
-      case "subscribers":
-        // In a real application, you'd want to restrict access to subscribers
-        // to authenticated admin users. For now, we fetch all.
-        data = await prisma.subscriber.findMany({
-          orderBy: { subscribed_at: "desc" },
-        });
-        break;
-      case "analytics":
-        // In a real application, you'd want to restrict access to analytics
-        // to authenticated admin users. For now, we fetch all.
-        data = await prisma.analytics.findMany({
-          orderBy: { date: "desc" },
         });
         break;
       default:
@@ -119,74 +153,23 @@ export  async function GET(
     }
 
     if (data === null) {
+      // This null check might not be strictly necessary if findMany always returns array
       return NextResponse.json(
-        { message: `${model} not found.` },
+        { message: `${model} not found or no data available.` },
         { status: 404 }
       );
     }
 
-    const processedData = Array.isArray(data)
-      ? data.map((item) => {
-          const newItem = { ...item };
-          if (
-            "features" in newItem &&
-            newItem.features &&
-            typeof newItem.features !== "string"
-          ) {
-            newItem.features = JSON.stringify(newItem.features);
-          }
-          if (
-            "technologies" in newItem &&
-            newItem.technologies &&
-            typeof newItem.technologies !== "string"
-          ) {
-            newItem.technologies = JSON.stringify(newItem.technologies);
-          }
-          if (
-            "gallery" in newItem &&
-            newItem.gallery &&
-            typeof newItem.gallery !== "string"
-          ) {
-            newItem.gallery = JSON.stringify(newItem.gallery);
-          }
-          if (
-            "tags" in newItem &&
-            newItem.tags &&
-            typeof newItem.tags !== "string"
-          ) {
-            newItem.tags = JSON.stringify(newItem.tags);
-          }
-          if (
-            "social_links" in newItem &&
-            newItem.social_links &&
-            typeof newItem.social_links !== "string"
-          ) {
-            newItem.social_links = JSON.stringify(newItem.social_links);
-          }
-          if (
-            "business_hours" in newItem &&
-            newItem.business_hours &&
-            typeof newItem.business_hours !== "string"
-          ) {
-            newItem.business_hours = JSON.stringify(newItem.business_hours);
-          }
-          if (
-            "skills" in newItem &&
-            newItem.skills &&
-            typeof newItem.skills !== "string"
-          ) {
-            newItem.skills = JSON.stringify(newItem.skills);
-          }
-          return newItem;
-        })
-      : data; 
-
-    return NextResponse.json(processedData, { status: 200 });
+    // `processedData` is likely not needed if Prisma is handling JSON fields correctly
+    // and you're not doing specific data transformations here.
+    return NextResponse.json(data, { status: 200 });
   } catch (error: any) {
     console.error("API Error:", error);
     return NextResponse.json(
       { message: "Internal Server Error", error: error.message },
       { status: 500 }
     );
+  } finally {
+    await prisma.$disconnect();
   }
 }
