@@ -1,20 +1,13 @@
 import type { Metadata } from "next";
 
 import "./globals.css";
-import ParallaxWrapper from "@/components/wrappers/ParallaxWrapper";
-import dynamic from "next/dynamic";
-import Chatbot from "@/components/Chatbot";
-import Navbar from "@/components/ui/Navbar";
-import Footer from "@/components/ui/Footer";
+import Navbar from "@/components/ui/navbar";
+import Footer from "@/components/ui/footer";
 
 export const metadata: Metadata = {
   title: "Info Desk",
-  description: "Maasai Inspired Infrmations Desk for Jiaminie.inc",
+  description: "Maasai Inspired Informations Desk for Jiaminie.inc",
 };
-
-const ClientLayout = dynamic(() => import("./_components/ClientLayout"), {
-  ssr: false,
-});
 
 export default function RootLayout({
   children,
@@ -24,14 +17,9 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body>
-        <ParallaxWrapper>
-          <ClientLayout>
-            <Navbar/>
-            {children}
-            <Chatbot />
-            <Footer/>
-          </ClientLayout>
-        </ParallaxWrapper>
+          <Navbar />
+          {children}
+          <Footer />
       </body>
     </html>
   );
