@@ -26,34 +26,33 @@ export default function Footer() {
     { name: "Contact", href: "/contact" },
   ];
   return (
-    <footer
-      className="relative bg-[var(--color-maasai-black)] text-white py-10 md:py-14 overflow-hidden bottom-0 w-full"
-    >
-      <div className="absolute inset-0 bg-transparent"></div>
+    <footer className="relative bg-black text-white py-10 md:py-14 overflow-hidden bottom-0 w-full">
+      <div className="absolute inset-0 bg-red-900/5"></div>
 
       <div className="container mx-auto px-6 relative z-10">
         <div className="grid md:grid-cols-4 gap-8 mb-8 md:mb-12">
           <div className="col-span-2">
             <div className="flex items-center gap-3 mb-4">
- <motion.div
-              className="relative"
-              whileHover={{ scale: 1.02 }}
-              transition={{ type: "spring", stiffness: 400, damping: 10 }}
-            >
-              <Link href="/" className="block">
-                <motion.span
-                  className={`text-xl sm:text-2xl font-bold tracking-tight
+              <motion.div
+                className="relative"
+                whileHover={{ scale: 1.02 }}
+                transition={{ type: "spring", stiffness: 400, damping: 10 }}
+              >
+                <Link href="/" className="block">
+                  <motion.span
+                    className={`text-xl sm:text-2xl font-bold tracking-tight
                   text-[var(--color-maasai-red)]
                   `}
-                  initial={{ opacity: 0 }}
-                  animate={{ opacity: 1 }}
-                  transition={{ delay: 0.2 }}
-                >
-                  Jiaminie
-                  <span className="text-sm font-normal opacity-80">.inc</span>
-                </motion.span>
-              </Link>
-            </motion.div>            </div>
+                    initial={{ opacity: 0 }}
+                    animate={{ opacity: 1 }}
+                    transition={{ delay: 0.2 }}
+                  >
+                    Jiaminie
+                    <span className="text-sm font-normal opacity-80">.inc</span>
+                  </motion.span>
+                </Link>
+              </motion.div>{" "}
+            </div>
             <p className="text-[var(--color-maasai-light-grey)] leading-relaxed mb-6 max-w-md">
               Empowering businesses through innovative digital solutions and
               sustainable technology practices.
@@ -99,7 +98,10 @@ export default function Footer() {
             <div className="space-y-4">
               <div className="flex items-center gap-3 text-[var(--color-maasai-light-grey)]">
                 <div className="w-8 h-8 bg-[var(--color-maasai-red)]/20 rounded-lg flex items-center justify-center">
-                  <Mail size={16} className="text-[var(--color-maasai-accent)]" />
+                  <Mail
+                    size={16}
+                    className="text-[var(--color-maasai-accent)]"
+                  />
                 </div>
                 <a
                   href="mailto:info@jiaminie.inc"
@@ -110,7 +112,10 @@ export default function Footer() {
               </div>
               <div className="flex items-center gap-3 text-[var(--color-maasai-light-grey)]">
                 <div className="w-8 h-8 bg-[var(--color-maasai-red)]/20 rounded-lg flex items-center justify-center">
-                  <Phone size={16} className="text-[var(--color-maasai-accent)]" />
+                  <Phone
+                    size={16}
+                    className="text-[var(--color-maasai-accent)]"
+                  />
                 </div>
                 <a
                   href="tel:+ +254-114-102-5756"
@@ -121,7 +126,10 @@ export default function Footer() {
               </div>
               <div className="flex items-center gap-3 text-[var(--color-maasai-light-grey)]">
                 <div className="w-8 h-8 bg-[var(--color-maasai-red)]/20 rounded-lg flex items-center justify-center">
-                  <MapPin size={16} className="text-[var(--color-maasai-accent)]" />
+                  <MapPin
+                    size={16}
+                    className="text-[var(--color-maasai-accent)]"
+                  />
                 </div>
                 <span>Nairobi, Kenya</span>
               </div>
@@ -129,7 +137,7 @@ export default function Footer() {
           </div>
         </div>
 
-       <div className="border-t border-[var(--color-maasai-dark-grey)] pt-6 md:pt-8 text-center">
+        <div className="border-t border-[var(--color-maasai-dark-grey)] pt-6 md:pt-8 text-center">
           <p className="text-[var(--color-maasai-light-grey)] flex items-center justify-center gap-2">
             © 2025 Jiaminie.Inc. All rights reserved.
           </p>
